requirements-dev.txt
pytest==5.3.*
pytest-benchmark==3.2.*
maturin==0.8.*
eth-utils==1.9.*
py-evm==0.3.0a19